+++ internal/interfaces/types.go
@@ -1,7 +1,17 @@
 package controller
 
+import inferno "github.com/llm-inferno/optimizer/pkg/core"
+
 // Captures response from ModelAnalyzer(s) per model
 type ModelAnalyzeResponse struct {
+	// feasible allocations for all accelerators
+	Allocations map[string]*ModelAcceleratorAllocation // accelerator name -> allocation
+}
+
+// Allocation details of an accelerator to a variant
+type ModelAcceleratorAllocation struct {
+	Allocation *inferno.Allocation // allocation result of model analyzer
+
 	RequiredPrefillQPS float64
 	RequiredDecodeQPS  float64
 	Reason             string
@@ -11,4 +21,16 @@ type ModelAnalyzeResponse struct {
 // ModelAnalyzer or Optimizer could consume
 type MetricsSnapshot struct {
 	ActualQPS float64
+}
+
+type ServiceClassEntry struct {
+	Model  string `yaml:"model"`
+	SLOITL int    `yaml:"slo-itl"`
+	SLOTTW int    `yaml:"slo-ttw"`
+}
+
+type ServiceClass struct {
+	Name     string              `yaml:"name"`
+	Priority int                 `yaml:"priority"`
+	Data     []ServiceClassEntry `yaml:"data"`
 }
