+++ internal/optimizer/optimizer.go
@@ -1,0 +1,49 @@
+package optimizer
+
+import (
+	"context"
+	"fmt"
+
+	llmdOptv1alpha1 "github.com/llm-d-incubation/inferno-autoscaler/api/v1alpha1"
+	interfaces "github.com/llm-d-incubation/inferno-autoscaler/internal/interfaces"
+	"github.com/llm-d-incubation/inferno-autoscaler/internal/utils"
+	inferno "github.com/llm-inferno/optimizer/pkg/core"
+	infernoManager "github.com/llm-inferno/optimizer/pkg/manager"
+)
+
+type VariantAutoscalingsEngine struct {
+	manager *infernoManager.Manager
+	system  *inferno.System
+}
+
+func NewVariantAutoscalingsEngine(manager *infernoManager.Manager, system *inferno.System) *VariantAutoscalingsEngine {
+	return &VariantAutoscalingsEngine{
+		manager: manager,
+		system:  system,
+	}
+}
+
+func (engine *VariantAutoscalingsEngine) Optimize(ctx context.Context,
+	vaList llmdOptv1alpha1.VariantAutoscalingList,
+	analysis map[string]*interfaces.ModelAnalyzeResponse,
+	metrics map[string]*interfaces.MetricsSnapshot,
+) (map[string]llmdOptv1alpha1.OptimizedAlloc, error) {
+
+	if err := engine.manager.Optimize(); err != nil {
+		return nil, err
+	}
+	allocationSolution := engine.system.GenerateSolution()
+	if allocationSolution == nil || len(allocationSolution.Spec) == 0 {
+		return nil, fmt.Errorf("no feasible solution found")
+	}
+
+	optimizedAllocMap := make(map[string]llmdOptv1alpha1.OptimizedAlloc)
+	for _, va := range vaList.Items {
+		vaName := va.Name
+		vaNamespace := va.Namespace
+		if optimizedAllocation, err := utils.SetDesiredAllocation(vaName, vaNamespace, allocationSolution); err == nil {
+			optimizedAllocMap[vaName] = *optimizedAllocation
+		}
+	}
+	return optimizedAllocMap, nil
+}
